+++ mrv2/lib/mrvPanels/mrvPanelWidget.h
@@ -51,9 +51,11 @@ namespace mrv
             void clear_controls();
             void refresh();
 
-            math::Box2i box() const;
-
+            //! Return the group box in global coordinates.
+            math::Box2i global_box() const;
+            
             bool is_panel() const { return g->docked(); };
+            
             virtual void save();
 
             virtual void dock();